複数のURLテキストから複数のContentレコードを作成
