+++ app/models/summary.rb
@@ -1,5 +1,5 @@
 class Summary < ApplicationRecord
   belongs_to :user
-  has_many :contents
+  has_many :contents, dependent: :destroy
   accepts_nested_attributes_for :contents
 end
